+++ src/main/java/frc/robot/RobotContainer.java
@@ -7,7 +7,7 @@ import frc.robot.Constants.AutoConstants;
 import frc.robot.Constants.DriveConstants;
 import frc.robot.Constants.FieldConstants;
 import frc.robot.Constants.TurnConstants;
-import frc.robot.commands.DriveCommandFactory;
+import frc.robot.commands.AutonomousCommandFactory;
 import frc.robot.subsystems.DrivetrainSubsystem;
 import frc.robot.utils.Controller;
 import edu.wpi.first.wpilibj.GenericHID;
@@ -67,7 +67,8 @@ public class RobotContainer {
     // Configure the button bindings
     configureButtonBindings();
 
-    DriveCommandFactory.swapAutonomousCommands();
+    //initialize Shuffleboard swapping of autonomous commands
+    AutonomousCommandFactory.swapAutonomousCommands();
   }
 
   /**
